// src/protocol/generated.ts
export interface CommandJsonProgram {
	settings: CommandJsonProgramSettings
	sessions: CommandJsonSession[]
	user_id?: string | null
	program_id: string
	program_name: string
	messages?: CommandJsonMessage[]
	version: string
}

export interface CommandJsonMessage {
	is_flashing?: boolean
	is_showing?: boolean
	id?: string
	version?: string
	content: string
}

export interface CommandJsonProgramSettings {
	flash_start_time_seconds?: number
	flash_length_seconds?: number
}

export interface CommandJsonSession {
	duration_seconds?: number | null
	mode?: string
	started_at_timestamp?: number | null
	notes?: string
	offset?: number | null
	flash_start_time_seconds?: number
	flash_length_seconds?: number
	name: string
	id?: string
	presenter_name?: string | null
	version?: string
}

export interface CommandPayloads {
	add_message: {
		content: string
	}
	add_session: {
		mode?: string
		seconds?: number | null
		notes?: string | null
		offset?: number | null
		flash_start_time_seconds?: number | null
		flash_length_seconds?: number | null
		name?: string | null
		session_id?: string | null
		presenter_name?: string | null
	}
	add_time: {
		seconds?: number | null
		ms?: number | null
	}
	delete_all_messages: {}
	delete_all_sessions: {}
	delete_message: {
		message_id: string
	}
	delete_program: {
		program_id: string
	}
	delete_session: {
		session_id: string
	}
	get_program: {}
	get_program_list: {}
	hide_message: {}
	move_message_down: {
		message_id: string
	}
	move_message_up: {
		message_id: string
	}
	move_session_down: {
		session_id: string
	}
	move_session_up: {
		session_id: string
	}
	navigate_next_session: {}
	navigate_previous_session: {}
	pause_session: {}
	request_control_center_status: {}
	request_messages_status: {}
	request_sessions_status: {}
	request_settings_status: {}
	request_status: {
		detailed?: boolean
	}
	request_view_status: {}
	reset_current_session: {}
	reset_program_elapsed_timer: {}
	send_program: {
		program: CommandJsonProgram
	}
	set_blackout: {
		action: string
	}
	set_brightness: {
		brightness: number
	}
	set_flash: {
		action: string
	}
	set_glow: {
		enabled: boolean
	}
	set_is_time_up_display: {
		enabled: boolean
	}
	show_idle: {}
	show_message: {
		message_id?: string | null
		text?: string | null
		flash?: boolean
	}
	show_session: {
		session_id: string
	}
	start_session: {
		session_id?: string | null
	}
	subtract_time: {
		seconds?: number | null
		ms?: number | null
	}
	toggle_flash: {}
	toggle_glow: {}
	toggle_is_time_up_display: {}
	update_message: {
		message_id: string
		content: string
	}
	update_session: {
		mode?: string | null
		seconds?: number | null
		notes?: string | null
		offset?: number | null
		flash_start_time_seconds?: number | null
		flash_length_seconds?: number | null
		name?: string | null
		session_id: string
		presenter_name?: string | null
	}
}